# apps/docs-site/static/openapi.yaml
openapi: 3.1.0
info:
  title: Spiritual Q&A Backend
  version: 0.1.0
paths:
...
  /api/chat/query:
    post:
      tags:
      - Chat
      summary: Chat Query
      description: 'Send a query to the RAG pipeline.


        Supports both guest (rate-limited, no persistence) and

        authenticated (with conversation history) modes.'
      operationId: chat_query_api_chat_query_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/conversations:
    get:
      tags:
      - Chat
      summary: List Conversations
      description: List all conversations for the authenticated user.
      operationId: list_conversations_api_chat_conversations_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationSummary'
                title: Response List Conversations Api Chat Conversations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/conversations/{conversation_id}:
    get:
      tags:
      - Chat
      summary: Load Conversation
      description: Load all messages in a conversation.
      operationId: load_conversation_api_chat_conversations__conversation_id__get
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/app__schemas__chat_schemas__MessageResponse'
                title: Response Load Conversation Api Chat Conversations  Conversation
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Chat
      summary: Delete Conversation
      description: Delete a conversation and all its messages.
      operationId: delete_conversation_api_chat_conversations__conversation_id__delete
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__schemas__common_schemas__MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat/conversations/{conversation_id}/export:
    post:
      tags:
      - Chat
      summary: Export Conversation
      description: Export a conversation as formatted Markdown.
      operationId: export_conversation_api_chat_conversations__conversation_id__export_post
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CitationResponse:
      properties:
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        page:
          anyOf:
          - type: integer
          - type: 'null'
          title: Page
        paragraph_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Paragraph Id
        relevance_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Relevance Score
        passage_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Passage Text
      type: object
      title: CitationResponse
      description: Citation referencing a source document passage.
...
    ConversationSummary:
      properties:
        id:
          type: string
          title: Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        message_count:
          type: integer
          title: Message Count
          default: 0
        last_message_preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Message Preview
      type: object
      required:
      - id
      - created_at
      - updated_at
      title: ConversationSummary
      description: Lightweight conversation metadata for listing.
    ExportResponse:
      properties:
        export_data:
          type: string
          title: Export Data
      type: object
      required:
      - export_data
      title: ExportResponse
      description: Exported conversation data.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QueryMetadata:
      properties:
        retrieval_time_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Retrieval Time Ms
        llm_time_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Llm Time Ms
        total_chunks_retrieved:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Chunks Retrieved
      type: object
      title: QueryMetadata
      description: Performance/diagnostic metadata from the RAG pipeline.
    QueryRequest:
      properties:
        query:
          type: string
          maxLength: 2000
          minLength: 1
          title: Query
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
          description: UUID of existing conversation (auth users)
        guest_session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Session Id
          description: UUID for guest rate-limit tracking
      type: object
      required:
      - query
      title: QueryRequest
      description: Chat query request. Supports guest and authenticated modes.
    QueryResponse:
      properties:
        answer:
          type: string
          title: Answer
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
        citations:
          items:
            $ref: '#/components/schemas/CitationResponse'
          type: array
          title: Citations
        metadata:
          anyOf:
          - $ref: '#/components/schemas/QueryMetadata'
          - type: 'null'
        guest_queries_remaining:
          anyOf:
          - type: integer
          - type: 'null'
          title: Guest Queries Remaining
      type: object
      required:
      - answer
      title: QueryResponse
      description: Chat query response with answer, citations, and metadata.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    app__schemas__chat_schemas__MessageResponse:
      properties:
        id:
          type: string
          title: Id
        sender:
          type: string
          title: Sender
        content:
          type: string
          title: Content
        citations:
          items:
            $ref: '#/components/schemas/CitationResponse'
          type: array
          title: Citations
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - id
      - sender
      - content
      - timestamp
      title: MessageResponse
      description: A single message in a conversation.
    app__schemas__common_schemas__MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
      description: Simple message response for operations like logout.